Restructure the project
- Move FXMLController class to controller package
- Rename DriveQuickstart to DriveController and move to drive package
- Change drive scope
- Create Constants class in utils package and add few constants
