+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
     id 'org.openjfx.javafxplugin' version '0.0.10'
 }
 
-mainClassName = 'com.hyperdrive.DriveQuickstart'
+mainClassName = 'com.hyperdrive.drive.DriveController'
 sourceCompatibility = 16
 targetCompatibility = 16
