+++ src/main/java/com/hyperdrive/controllers/FXMLController.java
@@ -1,4 +1,4 @@
-package com.hyperdrive;
+package com.hyperdrive.controllers;
 
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
