+++ src/main/java/com/hyperdrive/utils/Constants.java
@@ -1,0 +1,7 @@
+package com.hyperdrive.utils;
+
+public class Constants {
+
+    public static final String APPLICATION_NAME = "Google Drive API JavaFX Client";
+    public static final String APP_TITLE = "HyperDrive";
+}
