+++ src/main/resources/scene.fxml
@@ -6,7 +6,7 @@
 <StackPane xmlns:fx="http://javafx.com/fxml/1" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity"
            minWidth="-Infinity"
            prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.171"
-           fx:controller="com.hyperdrive.FXMLController">
+           fx:controller="com.hyperdrive.controllers.FXMLController">
     <children>
         <Label fx:id="label" text="Label"/>
     </children>
